backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pdfminer.six==20221105
python-docx==0.8.11
scikit-learn==1.3.2
spacy==3.7.2 